separate viz gallery from the modal
